feat: cartItem id 찾는 함수 util 폴더로 분리
